# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "roslina.com.pl documentation",
        "description": "System zarządzania domową rozsadą",
        "contact": {
            "email": "grave432@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://roslina.poligon",
            "description": "roslina.poligon"
        }
    ],
    "paths": {
        "/auth": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Get logged in user",
                "description": "Logout user and destroy all user's tokens",
                "operationId": "f6eb7532cc86794e4c976cbe4a76e9e6",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/ApiResponse401"
                    }
                }
            },
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Login user",
                "description": "Login user with valid crednecials and return token",
                "operationId": "eecd875305a510812e3cfb3b2ecf4597",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "token": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/ApiResponse401"
                    },
                    "422": {
                        "$ref": "#/components/responses/ApiResponse422"
                    }
                }
            },
            "delete": {
                "tags": [
                    "auth"
                ],
                "summary": "Logout user",
                "description": "Logout user and destroy all user's tokens",
                "operationId": "df2eee26444444a49a587c8073a4bf74",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "auth.logout_success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/packets/:id": {
            "get": {
                "tags": [
                    "packets"
                ],
                "summary": "Retrieve packet information",
                "description": "Get packet information with additions",
                "operationId": "9b5ba23c30919adbb6639c12f1adef3b",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "packet": {
                                            "$ref": "#/components/schemas/Packet"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/ApiResponse401"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/search/{search}": {
            "get": {
                "tags": [
                    "search"
                ],
                "summary": "A list of found items",
                "description": "A list of found items",
                "operationId": "07cb0aaffde0be9ebdbf265b5d2ab11c",
                "parameters": [
                    {
                        "name": "search",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "packets": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Packet"
                                                    },
                                                    {
                                                        "properties": {
                                                            "front": {
                                                                "$ref": "#/components/schemas/Image"
                                                            },
                                                            "back": {
                                                                "$ref": "#/components/schemas/Image"
                                                            }
                                                        }
                                                    }
                                                ]
                                            }
                                        },
                                        "producers": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/Producer"
                                                    }
                                                ]
                                            }
                                        },
                                        "users": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/User"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Image": {
                "properties": {
                    "id": {
                        "description": "image id",
                        "type": "integer"
                    },
                    "name": {
                        "description": "image description",
                        "type": "string"
                    },
                    "org_name": {
                        "description": "image orginalfile name",
                        "type": "string"
                    },
                    "file_name": {
                        "description": "image file name",
                        "type": "string"
                    },
                    "mime": {
                        "description": "image mime type",
                        "type": "string"
                    },
                    "size": {
                        "description": "image size",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "Packet": {
                "properties": {
                    "id": {
                        "description": "packet id",
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "name": {
                        "description": "packet name",
                        "type": "string"
                    },
                    "desc": {
                        "description": "packet description",
                        "type": "string"
                    },
                    "name_polish": {
                        "description": "polish package name",
                        "type": "string"
                    },
                    "name_latin": {
                        "description": "latin package name",
                        "type": "string"
                    },
                    "producer_id": {
                        "description": "package manufacturer id",
                        "type": "integer"
                    },
                    "expiration_date": {
                        "description": "package expiration date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "purchase_date": {
                        "description": "package purchase date",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "Producer": {
                "properties": {
                    "id": {
                        "description": "packet id",
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "name": {
                        "description": "pproducer name",
                        "type": "string"
                    },
                    "desc": {
                        "description": "producer description",
                        "type": "string"
                    },
                    "country": {
                        "description": "country of manufacture",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "description": "user's id",
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "name": {
                        "description": "user's name",
                        "type": "string"
                    },
                    "email": {
                        "description": "user's email",
                        "type": "string",
                        "format": "email"
                    },
                    "email_verified_at": {
                        "description": "date of email verification by the user",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true",
                        "nullable": "true"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            }
        },
        "responses": {
            "ApiResponse401": {
                "description": "Invalid credencials",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "code": {
                                    "type": "string",
                                    "example": "auth.invalid_credencials"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "ApiResponse422": {
                "description": "No required fields",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                },
                                "errors": {
                                    "properties": {
                                        "email": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The email field is required."
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    }
}